+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - [#495](https://github.com/teambit/bit/issues/495) support adding files to imported components and ignoring existing files
+- when a nested dependency is imported directly, re-link all its dependents
 
 ## [0.12.0-ext.11] - 2018-01-02